Extract test query implementations to `Given`
